chore(nx): upgrade to nx 16.10.0 (@nrwl -> @nx rename, workspace generators converted to local plugin)
