+++ webstack/apps/homebase-files/webpack.config.js
@@ -1,12 +1,10 @@
+const { composePlugins, withNx } = require('@nx/webpack');
 
-        const { composePlugins, withNx } = require('@nrwl/webpack');
-      
-        // Nx plugins for webpack.
-        module.exports = composePlugins(withNx(), (config) => {
-          // Update the webpack config as needed here.
-          // e.g. config.plugins.push(new MyPlugin())
-          // For more information on webpack config and Nx see:
-          // https://nx.dev/packages/webpack/documents/webpack-config-setup
-          return config;
-        });
-        
+// Nx plugins for webpack.
+module.exports = composePlugins(withNx(), (config) => {
+  // Update the webpack config as needed here.
+  // e.g. config.plugins.push(new MyPlugin())
+  // For more information on webpack config and Nx see:
+  // https://nx.dev/packages/webpack/documents/webpack-config-setup
+  return config;
+});
