+++ webstack/apps/webapp/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["plugin:@nrwl/nx/react", "..\\..\\.eslintrc.json"],
+  "extends": ["plugin:@nx/react", "..\\..\\.eslintrc.json"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
