+++ webstack/babel.config.json
@@ -1,4 +1,4 @@
 {
-  "presets": ["@nrwl/web/babel"],
+  "presets": ["@nx/web/babel"],
   "babelrcRoots": ["*"]
 }
