+++ webstack/libs/sagebase/.babelrc
@@ -1,7 +1,7 @@
 {
   "presets": [
     [
-      "@nrwl/js/babel",
+      "@nx/js/babel",
       {
         "useBuiltIns": "usage"
       }
