+++ webstack/libs/sagebase/project.json
@@ -5,29 +5,22 @@
   "projectType": "library",
   "targets": {
     "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": [
-        "{options.outputFile}"
-      ],
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": [
-          "libs/sagebase/**/*.ts"
-        ]
+        "lintFilePatterns": ["libs/sagebase/**/*.ts"]
       }
     },
     "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": [
-        "{options.outputPath}"
-      ],
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
       "options": {
         "outputPath": "dist/libs/sagebase",
         "tsConfig": "libs/sagebase/tsconfig.lib.json",
         "packageJson": "libs/sagebase/package.json",
         "main": "libs/sagebase/src/index.ts",
-        "assets": [
-          "libs/sagebase/*.md"
-        ]
+        "assets": ["libs/sagebase/*.md"],
+        "updateBuildableProjectDepsInPackageJson": true
       }
     }
   },
