+++ webstack/libs/shared/project.json
@@ -5,14 +5,10 @@
   "projectType": "library",
   "targets": {
     "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": [
-        "{options.outputFile}"
-      ],
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": [
-          "libs/shared/**/*.ts"
-        ]
+        "lintFilePatterns": ["libs/shared/**/*.ts"]
       }
     }
   },
