+++ webstack/package.json
@@ -25,8 +25,8 @@
     "prod-homebase": "rimraf dist && nx run-many --target=build --projects=homebase,webapp --configuration=production --verbose",
     "prod-files": "rimraf dist && nx run-many --target=build --projects=homebase-files --configuration=production --verbose",
     "prod-yjs": "rimraf dist && yarn build homebase-yjs --target=build --configuration=production --verbose",
-    "newapp": "nx workspace-generator --interactive newapp",
-    "regen": "nx workspace-generator regen"
+    "newapp": "nx g @sage3/workspace-plugin:newapp",
+    "regen": "nx g @sage3/workspace-plugin:regen"
   },
   "private": true,
   "dependencies": {
@@ -38,6 +38,7 @@
     "@monaco-editor/react": "^4.7.0",
     "@react-three/fiber": "^8.14.1",
     "@remotion/media-parser": "^4.0.339",
+    "@swc/helpers": "~0.5.2",
     "@turf/bbox": "^7.2.0",
     "@turf/center": "^7.2.0",
     "ansi-to-react": "^6.1.6",
@@ -147,19 +148,16 @@
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-react": "^7.14.5",
     "@babel/preset-typescript": "^7.15.0",
-    "@nrwl/cli": "15.9.7",
-    "@nrwl/devkit": "15.9.7",
-    "@nrwl/eslint-plugin-nx": "15.9.7",
-    "@nrwl/express": "15.9.7",
-    "@nrwl/js": "15.9.7",
-    "@nrwl/linter": "15.9.7",
-    "@nrwl/node": "15.9.7",
-    "@nrwl/react": "15.9.7",
-    "@nrwl/web": "15.9.7",
-    "@nrwl/webpack": "15.9.7",
-    "@nrwl/workspace": "15.9.7",
+    "@nx/webpack": "16.10.0",
+    "@nx/devkit": "16.10.0",
+    "@nx/js": "16.10.0",
+    "@nx/plugin": "16.10.0",
+    "@nx/workspace": "16.10.0",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
-    "@svgr/webpack": "^6.1.2",
+    "@svgr/webpack": "8.1.0",
+    "@swc-node/register": "~1.6.7",
+    "@swc/cli": "~0.1.62",
+    "@swc/core": "~1.3.85",
     "@testing-library/react": "13.4.0",
     "@types/compression": "^1.7.0",
     "@types/cookie-parser": "^1.4.2",
@@ -218,19 +216,25 @@
     "node-addon-api": "^8.7.0",
     "node-gyp": "^12.2.0",
     "node-loader": "^1.0.2",
-    "nx": "15.9.7",
+    "nx": "16.10.0",
     "prettier": "^3.6.2",
     "react-refresh": "^0.14.0",
     "rimraf": "^3.0.2",
     "simple-zustand-devtools": "^1.1.0",
     "style-loader": "^3.3.0",
-    "stylus": "^0.55.0",
+    "stylus": "0.59.0",
     "stylus-loader": "^7.1.0",
     "ts-node": "^10.9.2",
     "typescript": "5.5.2",
     "url-loader": "^4.1.1",
     "webpack": "^5.75.0",
-    "webpack-merge": "^5.8.0"
+    "webpack-merge": "^5.8.0",
+    "@nx/linter": "16.10.0",
+    "@nx/web": "16.10.0",
+    "@nx/eslint-plugin": "16.10.0",
+    "@nx/react": "16.10.0",
+    "@nx/express": "16.10.0",
+    "@nx/node": "16.10.0"
   },
   "engines": {
     "node": "^20 || ^22 || ^24"
