+++ webstack/tools/workspace-plugin/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@sage3/workspace-plugin",
+  "version": "0.0.1",
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "type": "commonjs",
+  "main": "./src/index.js",
+  "typings": "./src/index.d.ts",
+  "generators": "./generators.json"
+}
