+++ webstack/tools/workspace-plugin/project.json
@@ -1,0 +1,55 @@
+{
+  "name": "workspace-plugin",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "tools/workspace-plugin/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": [
+        "{options.outputPath}"
+      ],
+      "options": {
+        "outputPath": "dist/tools/workspace-plugin",
+        "main": "tools/workspace-plugin/src/index.ts",
+        "tsConfig": "tools/workspace-plugin/tsconfig.lib.json",
+        "assets": [
+          {
+            "input": "./tools/workspace-plugin/src",
+            "glob": "**/!(*.ts)",
+            "output": "./src"
+          },
+          {
+            "input": "./tools/workspace-plugin/src",
+            "glob": "**/*.d.ts",
+            "output": "./src"
+          },
+          {
+            "input": "./tools/workspace-plugin",
+            "glob": "generators.json",
+            "output": "."
+          },
+          {
+            "input": "./tools/workspace-plugin",
+            "glob": "executors.json",
+            "output": "."
+          }
+        ]
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": [
+        "{options.outputFile}"
+      ],
+      "options": {
+        "lintFilePatterns": [
+          "tools/workspace-plugin/**/*.ts",
+          "tools/workspace-plugin/package.json",
+          "tools/workspace-plugin/generators.json"
+        ]
+      }
+    }
+  },
+  "tags": []
+}
