+++ webstack/tools/workspace-plugin/src/generators/newapp/README.md
