+++ webstack/tools/workspace-plugin/src/generators/newapp/files/__name__.tsx__tmpl__
