+++ webstack/tools/workspace-plugin/src/generators/newapp/files/index.ts__tmpl__
