+++ webstack/tools/workspace-plugin/src/generators/newapp/files/styling.css__tmpl__
