+++ webstack/tools/workspace-plugin/src/generators/newapp/schema.json
