+++ webstack/tools/workspace-plugin/src/generators/regen/README.md
