+++ webstack/tools/workspace-plugin/src/generators/regen/schema.json
@@ -2,5 +2,5 @@
   "cli": "nx",
   "id": "regen",
   "type": "object",
-  "properties": { }
+  "properties": {}
 }
