+++ webstack/tsconfig.base.json
@@ -16,53 +16,23 @@
     "strictPropertyInitialization": true,
     "target": "es2015",
     "module": "esnext",
-    "typeRoots": [
-      "node_modules/@types",
-      "libs/shared/src/lib/types"
-    ],
-    "lib": [
-      "es2017",
-      "dom",
-      "es2018",
-      "es2019",
-      "es2020",
-      "es2021"
-    ],
+    "typeRoots": ["node_modules/@types", "libs/shared/src/lib/types"],
+    "lib": ["es2017", "dom", "es2018", "es2019", "es2020", "es2021"],
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@sage3/applications/apps": [
-        "libs/applications/src/lib/apps.ts"
-      ],
-      "@sage3/applications/initialValues": [
-        "libs/applications/src/lib/initialValues.ts"
-      ],
-      "@sage3/applications/schema": [
-        "libs/applications/src/lib/schema.ts"
-      ],
-      "@sage3/backend": [
-        "libs/backend/src/index.ts"
-      ],
-      "@sage3/frontend": [
-        "libs/frontend/src/index.ts"
-      ],
-      "@sage3/sagebase": [
-        "libs/sagebase/src/index.ts"
-      ],
-      "@sage3/sageplugin": [
-        "libs/sageplugin/src/index.ts"
-      ],
-      "@sage3/shared": [
-        "libs/shared/src/index.ts"
-      ],
-      "@sage3/shared/types": [
-        "libs/shared/src/lib/types/index.ts"
-      ]
+      "@sage3/applications/apps": ["libs/applications/src/lib/apps.ts"],
+      "@sage3/applications/initialValues": ["libs/applications/src/lib/initialValues.ts"],
+      "@sage3/applications/schema": ["libs/applications/src/lib/schema.ts"],
+      "@sage3/backend": ["libs/backend/src/index.ts"],
+      "@sage3/frontend": ["libs/frontend/src/index.ts"],
+      "@sage3/sagebase": ["libs/sagebase/src/index.ts"],
+      "@sage3/sageplugin": ["libs/sageplugin/src/index.ts"],
+      "@sage3/shared": ["libs/shared/src/index.ts"],
+      "@sage3/shared/types": ["libs/shared/src/lib/types/index.ts"],
+      "@sage3/workspace-plugin": ["tools/workspace-plugin/src/index.ts"]
     }
   },
-  "exclude": [
-    "node_modules",
-    "tmp"
-  ]
+  "exclude": ["node_modules", "tmp"]
 }
